feat(delete books): user should be able to delete book
- Consume the `delete` book endpoint

[Delivers]
