+++ src/store/auths/actions.js
@@ -1,53 +1,8 @@
-/* eslint no-console: 0 */
-import server from '../../server.config';
-import { decodeToken } from '../Utils/token';
-import router from '../routes';
-import toast from '../Utils/toasts';
-import apiErrorHandler from '../Utils/apiErrorHandler';
-
-
-const state = {
-    user: {
-        details: null,
-        loggedIn: false
-    },
-    message: '',
-    loading: false,
-    success: false,
-    errors: null
-};
-
-const mutations = {
-    postUser: function(state) {
-        state.loading = true;
-    },
-    postUserSuccess: function(state, response) {
-        const { userDetails, success, message } = response;
-
-        state.user.details = userDetails;
-        state.success = success;
-        state.message = message;
-        state.user.loggedIn = true;
-        state.loading = false;
-        state.errors = null;
-    },
-    postUserFailure: function(state, error) {
-        state.errors = error;
-        state.loading = false;
-        state.success = false;
-        state.message = ''
-    },
-    logOutUser: function(state) {
-        state.user = {
-            details: null,
-            loggedIn: false
-        };
-        state.message = '';
-        state.loading = false;
-        state.success = false;
-        state.errors = null;
-    }
-};
+import server from '../../../server.config';
+import { decodeToken } from '../../Utils/token';
+import router from '../../routes';
+import toast from '../../Utils/toasts';
+import apiErrorHandler from '../../Utils/apiErrorHandler';
 
 const actions = {
     signUserIn: async function({ commit }, 
@@ -92,11 +47,4 @@ const actions = {
     }
 };
 
-
-
-export default {
-    namespaced: true,
-    state,
-    mutations,
-    actions
-};
+export default actions;
