+++ src/views/Dashboard/Favourite/Favourite.vue
@@ -5,6 +5,7 @@
         @modalToOpen="handleModalEvent"
         noBooksMessage="No favourite book yet!"
         :showSpinner="bookLoading"
+        :userId="userInfo.id"
     />
 </template>
